Take precompile part out
